Move new DataHelper Exceptions into seperate folder
