+++ src/FINDOLOGIC/Export/Exceptions/AttributeKeyLengthException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class AttributeKeyLengthException extends InternalCharacterLimitException
+{
+    public function __construct($attributeKey, $characterLimit)
+    {
+        parent::__construct('Attribute with name', $attributeKey, $characterLimit);
+    }
+}
