+++ src/FINDOLOGIC/Export/Exceptions/AttributeValueLengthException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class AttributeValueLengthException extends InternalCharacterLimitException
+{
+    public function __construct($attributeName, $characterLimit)
+    {
+        parent::__construct('Value of attribute', $attributeName, $characterLimit);
+    }
+}
